requirements.txt
mysql-connector-python==9.0.0
pandas==1.5.3